(#34) refactor: 컴퍼넌트로 분리
