+++ src/components/auth/findAccount/Certification.tsx
@@ -16,6 +16,8 @@ import {
 } from '@/apis/auth';
 import { useErrorMessage } from '@/hooks/useErrorMessage';
 import { useToast } from '@/hooks/useToast';
+import { Email } from '@/components/auth/findAccount/Email';
+import { AuthCode } from '@/components/auth/findAccount/AuthCode';
 
 interface Props {
   account_id: string;
@@ -39,7 +41,7 @@ export function Certification({
   const { errorMessages, changeErrorMessage } = useErrorMessage(errorTypes);
   const { toastDispatch } = useToast();
   const [emailHint, setEmailHint] = useState('');
-  const onClickCheckEmail = () => {
+  const checkEmail = () => {
     checkEmailDuplicate(account_id)
       .then((res) => {
         setStep('EMAIL');
@@ -55,7 +57,7 @@ export function Certification({
         }
       });
   };
-  const onClickPostEmailAuthCode = (type?: 'resend') => {
+  const postEmailCode = (type?: 'resend') => {
     if (!email) {
       changeErrorMessage('email', '이메일을 입력해 주세요.');
       return;
@@ -85,7 +87,7 @@ export function Certification({
         }
       });
   };
-  const onClickCheckEmailAuthCode = () => {
+  const checkEmailCode = () => {
     checkEmailAuthCode(email, auth_code, 'PASSWORD')
       .then(() => {
         setStep('RESET');
@@ -100,6 +102,11 @@ export function Certification({
         changeErrorMessage('auth_code', '인증코드가 일치하지 않습니다.');
       });
   };
+  const verificationBtn = () => {
+    if (step === 'ACCOUNT_ID') checkEmail();
+    else if (step === 'EMAIL') postEmailCode();
+    else checkEmailCode();
+  };
   return (
     <>
       <_AccountIdInput
@@ -113,61 +120,23 @@ export function Certification({
         errorMsg={errorMessages?.account_id}
       />
       {step !== 'RESET' && step !== 'ACCOUNT_ID' && (
-        <>
-          <_EmailHintBox>
-            <Text fontSize="s" display="block">
-              아이디와 일치하는 이메일입니다.
-            </Text>
-            <Text fontSize="s" color="primary">
-              {emailHint}
-            </Text>
-          </_EmailHintBox>
-          <_EmailInput
-            label="이메일"
-            placeholder="이메일을 입력해주세요"
-            onChange={onChangeValue}
-            width={480}
-            type="text"
-            name="email"
-            value={email}
-            errorMsg={errorMessages?.email}
-          />
-        </>
+        <Email
+          emailHint={emailHint}
+          onChangeValue={onChangeValue}
+          email={email}
+          errorMessage={errorMessages?.email}
+        />
       )}
       {step === 'AUTH_CODE' && (
-        <_AuthCodeArea>
-          <_AuthCodeInput
-            label="인증코드"
-            name="auth_code"
-            onChange={onChangeValue}
-            type="text"
-            width={480}
-            placeholder="이메일로 발송된 인증코드를 입력해주세요."
-            value={auth_code}
-            errorMsg={errorMessages?.auth_code}
-          />
-          <_ReSendAuthCodeWrapper>
-            <_ResendQuestion fontSize="s" color="gray6">
-              인증번호가 발송되지 않았나요?
-            </_ResendQuestion>
-            <Button
-              onClick={() => onClickPostEmailAuthCode('resend')}
-              color="gray"
-              type="underline"
-              clickType="button"
-              size="default"
-            >
-              인증번호 재발송
-            </Button>
-          </_ReSendAuthCodeWrapper>
-        </_AuthCodeArea>
+        <AuthCode
+          auth_code={auth_code}
+          errorMessage={errorMessages?.auth_code || ''}
+          onClickPostEmailAuthCode={postEmailCode}
+          onChangeValue={onChangeValue}
+        />
       )}
       <_NextButton
-        onClick={() => {
-          if (step === 'ACCOUNT_ID') onClickCheckEmail();
-          else if (step === 'EMAIL') onClickPostEmailAuthCode();
-          else onClickCheckEmailAuthCode();
-        }}
+        onClick={verificationBtn}
         size="default"
         color="primary"
         type="contained"
@@ -180,45 +149,6 @@ export function Certification({
 
 const _AccountIdInput = styled(Input)`
   margin-top: 56px;
-  margin-bottom: 40px;
-`;
-
-const _EmailInput = styled(Input)`
-  margin-bottom: 40px;
-`;
-
-const _ResendQuestion = styled(Text)`
-  margin-right: 12px;
-`;
-
-const _EmailHintBox = styled.div`
-  margin: 16px 0 40px 0;
-  padding: 12px 16px;
-  width: 480px;
-  height: 70px;
-  background-color: ${({ theme }) => theme.color.gray2};
-  > p {
-    // 디자인 시스템 구축시 반영 예정
-    line-height: 24px;
-  }
-`;
-
-const _AuthCodeArea = styled.div`
-  position: relative;
-  margin-top: 60px;
-`;
-
-const _ReSendAuthCodeWrapper = styled.div`
-  display: flex;
-  position: absolute;
-  right: 0;
-  top: -5px;
-  > button {
-    padding: 0;
-  }
-`;
-
-const _AuthCodeInput = styled(Input)`
   margin-bottom: 40px;
 `;
